feat(Object): have a single IOTA-Names button for all networks
